created the ConfigFor... classes
